backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
openai>=1.6.1
pinecone>=6.0.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.0.20
pypdf2>=3.0.1
python-dotenv>=1.0.0
numpy>=1.24.3
tiktoken>=0.5.2
